Update stackInsert() in algo.c
Reverted deadend code to previous implementation, leaving comments for easy swap-out.
Modified stackInsert() to return a bool that indicated if the node was added to the stack. This allows code to do different things when revisiting a node.
Fixed last component of stackInsert() to correctly copy the temporary array into the nodeCurrent array.