Add files via upload
First implementation of a negotiator to exchange commands between the mms simulator running on a PC, and an arduino connected to that PC.
Whether any of this code actually does anything or not, I'm not sure; somebody with access to an arduino will need to test it once we've ported Mathazzar-patch-7 to it for testing processing time. Someone else who knows more about reading serial streams in C looking at this would be appreciated.
It's currently set up to expect an integer to be passed from the arduino, which it will interpret as a particular command, and execute that command in the simulator, then return the result unmodified back to the arduino.